updating test to match llama.cpp
