+++ .github/workflows/ci.yml
@@ -4,10 +4,17 @@ on:
   - pull_request
   - workflow_dispatch
 env:
-  REASONING_MODEL_URL: https://huggingface.co/unsloth/Phi-4-mini-instruct-GGUF/resolve/main/Phi-4-mini-instruct-Q2_K.gguf
-  REASONING_MODEL_NAME: Phi-4-mini-instruct-Q2_K.gguf
+
+  REASONING_MODEL_URL: https://huggingface.co/ggml-org/models/resolve/main/tinyllamas/stories260K.gguf
+  REASONING_MODEL_NAME: stories260K.gguf
+  INFILL_MODEL_URL: https://huggingface.co/ggml-org/models/resolve/main/tinyllamas/stories260K-infill.gguf
+  INFILL_MODEL_NAME: stories260K-infill.gguf
+  MOE_MODEL_URL: https://huggingface.co/ggml-org/stories15M_MOE/resolve/main/stories15M_MOE-F16.gguf
+  MOE_MODEL_NAME: stories15M_MOE-F16.gguf
   RERANKING_MODEL_URL: https://huggingface.co/gpustack/jina-reranker-v1-tiny-en-GGUF/resolve/main/jina-reranker-v1-tiny-en-Q4_0.gguf
   RERANKING_MODEL_NAME: jina-reranker-v1-tiny-en-Q4_0.gguf
+  EMBEDDING_MODEL_URL: https://huggingface.co/ggml-org/models/resolve/main/bert-bge-small/ggml-model-f16.gguf
+  EMBEDDING_MODEL_NAME: ggml-model-f16.gguf
 jobs:
 
   build-and-test-linux:
@@ -25,8 +32,18 @@ jobs:
           .github/build.sh -DLLAMA_VERBOSE=ON
       - name: Download reranking model 
         run: curl -L ${RERANKING_MODEL_URL} --create-dirs -o models/${RERANKING_MODEL_NAME}
+
       - name: Download reasoning calling model
         run: curl -L ${REASONING_MODEL_URL} --create-dirs -o models/${REASONING_MODEL_NAME}
+
+      - name: Download infill  calling model
+        run: curl -L ${INFILL_MODEL_URL} --create-dirs -o models/${INFILL_MODEL_NAME}
+
+      - name: Download MOE  model
+        run: curl -L ${MOE_MODEL_URL} --create-dirs -o models/${MOE_MODEL_NAME}
+
+      - name: Download EMBEDDING  model
+        run: curl -L ${EMBEDDING_MODEL_URL} --create-dirs -o models/${EMBEDDING_MODEL_NAME}
 
       - name: List files in models directory
         run: ls -l models/
@@ -60,10 +77,22 @@ jobs:
         run: |
           mvn compile
           .github/build.sh ${{ matrix.target.cmake }}
+
       - name: Download reranking model 
         run: curl -L ${RERANKING_MODEL_URL} --create-dirs -o models/${RERANKING_MODEL_NAME}
+
       - name: Download reasoning calling model
         run: curl -L ${REASONING_MODEL_URL} --create-dirs -o models/${REASONING_MODEL_NAME}
+      
+      - name: Download infill  calling model
+        run: curl -L ${INFILL_MODEL_URL} --create-dirs -o models/${INFILL_MODEL_NAME}
+
+      - name: Download MOE  model
+        run: curl -L ${MOE_MODEL_URL} --create-dirs -o models/${MOE_MODEL_NAME}
+
+      - name: Download EMBEDDING  model
+        run: curl -L ${EMBEDDING_MODEL_URL} --create-dirs -o models/${EMBEDDING_MODEL_NAME}
+
       - name: List files in models directory
         run: ls -l models/
       - name: Run tests
@@ -88,10 +117,22 @@ jobs:
         run: |
           mvn compile
           .github\build.bat -DLLAMA_VERBOSE=ON
+
       - name: Download reranking model 
         run: curl -L $env:RERANKING_MODEL_URL --create-dirs -o models/$env:RERANKING_MODEL_NAME
+
       - name: Download reasoning calling model
         run: curl -L $env:REASONING_MODEL_URL --create-dirs -o models/$env:REASONING_MODEL_NAME
+      
+      - name: Download infill  calling model
+        run: curl -L $env:INFILL_MODEL_URL --create-dirs -o models/$env:INFILL_MODEL_NAME
+
+      - name: Download MOE  model
+        run: curl -L $env:MOE_MODEL_URL --create-dirs -o models/$env:MOE_MODEL_NAME
+
+      - name: Download EMBEDDING  model
+        run: curl -L $env:EMBEDDING_MODEL_URL --create-dirs -o models/$env:EMBEDDING_MODEL_NAME
+
       - name: List files in models directory
         run: ls -l models/
       - name: Run tests
